Performance optimizations for javascript output
Since javascript doesn't have structs, all enumerators cause a memory
allocation.  Switching to for loops to avoid this memory allocation is a
big win for javascript, and should be negligable change for normal .net
execution.
Also changed colorSetsSeen to be a boolean lookup array rather than a
list with add/remove.  This should be a performance optimization in
general.